+++ app/Gateways/ProviderApiClient.php
@@ -12,8 +12,8 @@ namespace App\Gateways;
 
 use Psr\Http\Message\ResponseInterface;
 use WNowicki\Generic\ApiClient\AbstractApiClient;
-use WNowicki\Generic\ApiClient\BadResponseException;
 use WNowicki\Generic\ApiClient\ErrorResponseException;
+use WNowicki\Generic\ApiClient\WrongResponseException;
 
 /**
  * Class ProviderApiClient
@@ -56,15 +56,14 @@ class ProviderApiClient extends AbstractApiClient
      * @author WN
      * @param ResponseInterface $response
      * @return array
-     * @throws BadResponseException
+     * @throws WrongResponseException
      */
     protected function processResponse(ResponseInterface $response)
     {
-        // json_decode returns null if the json can not be decoded, but will also return null is the json is null
-        if (($responseBody = json_decode($response->getBody()->getContents(), true)) !== null) {
+        if ($responseBody = json_decode($response->getBody()->getContents(), true)) {
             return $responseBody;
         }
-        throw new BadResponseException('Response body was malformed JSON', $response->getStatusCode());
+        throw new WrongResponseException('Response body was malformed JSON', $response->getStatusCode());
     }
 
     /**