+++ app/User.php
@@ -17,10 +17,12 @@ use Zizaco\Entrust\Traits\EntrustUserTrait;
  * @property string $email
  * @property string $password
  * @property int    $merchant_id
- * @property string $locations
  * @property string $remember_token
  * @property        $created_at
  * @property        $updated_at
+ * @property Role[]             $roles
+ * @property Basket\Location[]  $locations
+ *
  * @package App
  */
 class User extends Model implements AuthenticatableContract, CanResetPasswordContract
@@ -61,6 +63,10 @@ class User extends Model implements AuthenticatableContract, CanResetPasswordCon
 
     /**
      * Get the role record associated with the user.
+     *
+     * TODO To remove! ~WN
+     *
+     * @deprecated
      * @author MS
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */