+++ resources/views/master.blade.php
@@ -108,5 +108,6 @@
         r.parentNode.insertBefore(e,r)}(window,document,'script','ga'));
     ga('create','UA-XXXXX-X','auto');ga('send','pageview');
 </script>
+@yield('scripts')
 </body>
 </html>