+++ resources/views/user/show.blade.php
@@ -7,7 +7,7 @@
 
     <h2>
         {{ Str::upper(' view ' . str_singular(Request::segment(1))) }}
-        @include('includes.page.show_details_button_group', ['id'=>$user->id,'edit'=>true,'delete'=>true])
+        @include('includes.page.show_details_button_group', ['id'=>$user->id,'edit'=>true,'delete'=>true, 'locations'=>true])
     </h2>
     @include('includes.page.breadcrumb', ['override2'=>$user->name])
 
@@ -42,10 +42,22 @@
                         </li>
                         <li class="list-group-item">
                             @if($user->locations !== null)
-                                <strong>Locations: </strong><br />
+                                <strong>Locations: </strong>
+                                <ul>
                                 @foreach ($user->locations as $location)
-                                    {{ $location->name }}<br />
+                                    <li>{{ $location->name }}</li>
                                 @endforeach
+                                </ul>
+                            @endif
+                        </li>
+                        <li class="list-group-item">
+                            @if($user->roles !== null)
+                                <strong>Roles: </strong>
+                                <ul>
+                                @foreach ($user->roles as $role)
+                                    <li>{{ $role->display_name }}</li>
+                                @endforeach
+                                </ul>
                             @endif
                         </li>
                     </ul>